Add files via upload, original project code
This is the original version of the puzzle slider game before I took Object Oriented Design. This legacy code is no longer used in my reworked project version on the main branch. Compared to the new version, this project is done procedurally, where in the new version I use Object Oriented classes which resulted in much cleaner code, and reduced the amount of lines I used drastically. Open the PS_Rework file to look at the new implementation of this project.